==> scripts/02-download_data.R <==
#### Preamble ####
# Purpose: Downloading raw Trackinginjustice Website Data 
# Author: Yuanting Han
# Date:  29 November 2024
# Contact: q2900725357@gmail.com
# License: MIT
# Pre-requisites: 
  # - The raw source data Trackinginjustice Website Data Sheet.xlsx must be downloaded manually
# Any other information needed?  Make sure in the  `Injustice` rproj


#### Workspace setup ####
#no codes needed
#### Download data ####

#Note, a javascript is embeded in the webpage to control the downloading of the Trackinginjustice data
#thus, downloading directly using a link is not appropriate for getting data from this dynamic website page
#please, following the 4 steps as below:

#step 1. go to website  https://trackinginjustice.ca/explore-the-data
#step 2. scroll down the page until find the yellow colored "Download Excel" button
#step 3. push the button and got the "Trackinginjustice Website Data Sheet.xlsx"

#### Save data ####
#step 4. save the Trackinginjustice Website Data Sheet.xlsx into `data/01-raw_data`
